travel fix

+++ src/oceantrader/Travel.java
@@ -35,7 +35,7 @@ public class Travel {
                 String errorMsg = String.format("Not enough fuel!"
                                 + "\nYour ship only has %d fuel."
                                 + "\nTraveling to %s will cost %d fuel.",
-                        capacity, Map.selected.getName(), cost);
+                        capacity, Map.selected.getName(), (int) cost);
                 JOptionPane.showMessageDialog(window, errorMsg);
             } else {
                 int confirmPage = JOptionPane.showConfirmDialog(window,
